airflowdags/requirements.txt
beautifulsoup4==4.12.3
boto3==1.34.54
pandas==2.2.1
pyarrow==15.0.2
psycopg2-binary==2.9.9
sqlalchemy==1.4.46
astro-run-dag
apache-airflow-providers-postgres